Create README.md
Problem Statement: In situations where your inbound traffic is higher than can
reasonably be managed by human support staff, an AI chatbot is one of your best
resources to handle the higher volume. Chatbots can be programmed to respond to
FAQs in real time. Develop a chatbot for the same.
Task Instructions
● Install Tensorflow, Keras, Pickle, nltk.
● Import necessary packages and initialize the variables.
● The data file is named intents.json and is in JSON format so we used the JSON
package to parse the JSON file into Python.
● Create two empty and name them documents and classes.
● Write a function to iterate through the intent data and tokenize the sentence using
nltk.wordtokenize() function and append each word in the word list of intent along
with the tag in documents.
● Append tags in intends to the empty list named classes.
● lemmatize each word and remove duplicate words from the list and create a pickle
file to store the Python objects which we will use while predicting.
● documents = combination between patterns and intents, classes = intents, words =
all words, vocabulary.
● Save words and classes ans words.pkl and classes.pkl
Create the training data in which we will provide the input and the output. Input will
be the pattern and output will be the class our input pattern belongs to. But the
computer doesn’t understand text so we will convert text into numbers.
● Lemmatize each word - create base word, in attempt to represent related words
● Create our bag of words array with 1, if word match found in current pattern
● Output is a '0' for each tag and '1' for current tag (for each pattern)
● Create training set for X and Y
● Once the training data is ready, build a deep neural network . Use the Keras
sequential API for this. Save the model as ‘chatbot_model.h5’.
● Load the trained model and then use a graphical user interface that will predict the
response from the bot. The model will only tell the class it belongs to, so implement
some functions which will identify the class and then retrieve a random response
from the list of responses.
● To predict the sentences and get a response from the user to let us create a new file
‘chatapp.py’.
● To predict the class, we will need to provide input in the same way as we did while
training. So we will create some functions that will perform text preprocessing and
then predict the class.
● After predicting the class, we will get a random response from the list of intents.
● develop a graphical user interface. Let’s use the Tkinter library which is shipped with
tons of useful libraries for GUI. We will take the input message from the user and
then use the helper functions we have created to get the response from the bot and
display it on the GUI.
● To run the chatbot, we have two main files; train_chatbot.py and chatapp.py, If we
don’t see any error during training, we have successfully created the model. Then to
run the app, we run the second file. The program will open up a GUI window within a
few seconds. With the GUI you can easily chat with the bot.